parallelExport {parallelMap}	R Documentation

Export R objects for parallelization.

Description

Makes sure that the objects are exported to slave process so that they can be used in a job function which is later run with parallelMap.

Usage

parallelExport(..., objnames, master = FALSE, level = as.character(NA),
  show.info = NA)

Arguments

...
[character]
Names of objects to export.

objnames
[character(1)]
Names of objects to export. Alternative way to pass arguments.

level
[character(1)]
If a (non-missing) level is specified in parallelStart, the function only exports if the level specified here matches. See parallelMap. Useful if this function is used in a package. Default is NA.

master
[logical(1)]
Really export to .GlobalEnv on master for local and multicore mode? Default is FALSE.

show.info
[logical(1)]
Verbose output on console? Can be used to override setting from options / parallelStart. Default is NA which means no overriding.

Value

Nothing.

[Package parallelMap version 1.1 Index]