parallelLapply {parallelMap}	R Documentation

Parallel versions of apply-family functions.

Description

parallelLapply: A parallel lapply version.
parallelSapply: A parallel sapply version.
All functions are simple wrappers for parallelMap.

Usage

parallelLapply(xs, fun, ..., level = NA_character_)

parallelSapply(xs, fun, ..., simplify = TRUE, use.names = TRUE,
  level = NA_character_)

Arguments

xs
[vector | list]
fun is applied to the elements of this argument.

fun
[function]
Function to map over xs.

...
[any]
Further arguments passed to fun.

simplify
[logical(1)]
See sapply. Default is TRUE.

use.names
[logical(1)]
See sapply. Default is TRUE.

level
[character(1)]
See parallelMap.

Value

For parallelLapply an unamed list, for parallelSapply it depends on the return value of fun and the settings of simplify and use.names.

[Package parallelMap version 1.1 Index]