parallelLibrary {parallelMap}	R Documentation

Load packages for parallelization.

Description

Makes sure that the packages are loaded in slave process so that they can be used in a job function which is later run with parallelMap.

For all modes, the packages are also (potentially) loaded on the master.

Usage

parallelLibrary(..., packages, master = TRUE, level = as.character(NA),
  show.info = NA)

Arguments

...
[character]
Names of packages to load.

packages
[character(1)]
Names of packages to load. Alternative way to pass arguments.

master
[logical(1)]
Load packages also on master? Default is TRUE.

level
[character(1)]
If a (non-missing) level is specified in parallelStart, the function only loads the packages if the level specified here matches. See parallelMap. Useful if this function is used in a package. Default is NA.

show.info
[logical(1)]
Verbose output on console? Can be used to override setting from options / parallelStart. Default is NA which means no overriding.

Value

Nothing.

[Package parallelMap version 1.1 Index]