parallelShowRegisteredLevels {parallelMap}	R Documentation

Display registered parallelization levels for all currently loaded packages.

Description

Display registered parallelization levels for all currently loaded packages.

Usage

parallelShowRegisteredLevels()

Value

Invisibly returns a list object that contains the displayed information.

[Package parallelMap version 1.1 Index]