parallelSource {parallelMap}	R Documentation

Source R files for parallelization.

Description

Makes sure that the files are sourced in slave process so that they can be used in a job function which is later run with parallelMap.

For all modes, the files are also (potentially) loaded on the master.

Usage

parallelSource(..., files, master = TRUE, level = as.character(NA),
  show.info = NA)

Arguments

...
[character]
File paths to sources.

files
[character]
File paths to sources. Alternative way to pass arguments.

master
[logical(1)]
Source files also on master for any mode? Default is TRUE.

level
[character(1)]
If a (non-missing) level is specified in parallelStart, the function only sources the files if the level specified here matches. See parallelMap. Useful if this function is used in a package. Default is NA.

show.info
[logical(1)]
Verbose output on console? Can be used to override setting from options / parallelStart. Default is NA which means no overriding.

Value

Nothing.

[Package parallelMap version 1.1 Index]