parallelStart {parallelMap}	R Documentation

Parallelization setup for parallelMap.

Description

Defines the underlying parallelization mode for parallelMap. Also allows to set a “level” of parallelization. Only calls to parallelMap with a matching level are parallelized. The defaults of all settings are taken from your options, which you can also define in your R profile. For an introductory tutorial and information on the options configuration, please go to the project's github page at https://github.com/berndbischl/parallelMap.

Usage

parallelStart(mode, cpus, socket.hosts, bj.resources = list(), logging,
  storagedir, level, show.info, ...)

parallelStartLocal(show.info)

parallelStartMulticore(cpus, logging, storagedir, level, show.info)

parallelStartSocket(cpus, socket.hosts, logging, storagedir, level, show.info,
  ...)

parallelStartMPI(cpus, logging, storagedir, level, show.info, ...)

parallelStartBatchJobs(bj.resources = list(), logging, storagedir, level,
  show.info)

Arguments

mode
[character(1)]
Which parallel mode should be used: “local”, “multicore”, “socket”, “mpi”, “BatchJobs”. Default is the option parallelMap.default.mode or, if not set, “local” without parallel execution.

cpus
[integer(1)]
Number of used cpus. For local and BatchJobs mode this argument is ignored. For socket mode, this is the number of processes spawned on localhost, if you want processes on multiple machines use socket.hosts. Default is the option parallelMap.default.cpus or, if not set, detectCores for multicore mode, mpi.universe.size for mpi mode and 1 for socket mode.

socket.hosts
[character]
Only used in socket mode, otherwise ignored. Names of hosts where parallel processes are spawned. Default is the option parallelMap.default.socket.hosts, if this option exists.

bj.resources
[list]
Resources like walltime for submitting jobs on HPC clusters via BatchJobs. See submitJobs. Defaults are taken from your BatchJobs config file.

logging
[logical(1)]
Should slave output be logged to files via sink under the storagedir? Files are named "<iteration_number>.log" and put into unique subdirectories named “parallelMap_log_<nr>” for each subsequent parallelMap operation. Previous logging directories are removed on parallelStart if logging is enabled. Logging is not supported for local mode, because you will see all output on the master and can also run stuff like traceback in case of errors. Default is the option parallelMap.default.logging or, if not set, FALSE.

storagedir
[character(1)]
Existing directory where log files and intermediate objects for BatchsJobs mode are stored. Note that all nodes must have write access to exactly this path. Default is the current working directory.

level
[character(1)]
You can set this so only calls to parallelMap are parallelized that have the same level specified. Default is the option parallelMap.default.level or, if not set, NA which means all calls to parallelMap are are parallelized.

show.info
[logical(1)]
Verbose output on console for all further package calls? Default is the option parallelMap.default.show.info or, if not set, TRUE.

...
[any]
Optional parameters, for socket mode passed to makePSOCKcluster, for mpi mode passed to makeCluster.

Details

Currently the following modes are supported, which internally dispatch the mapping operation to functions from different parallelization packages:

local

No parallelization with mapply.

multicore

Multicore execution on a single machine withmcmapply.

socket

Socket cluster on one or multiple machines with makePSOCKcluster and clusterMap.

mpi

Snow MPI cluster on one or multiple machines with makeCluster and clusterMap.

BatchJobs

Parallelization on batch queuing HPC clusters, e.g., Torque, SLURM, etc., with batchMap.

For BatchJobs mode you need to define a storage directory through the argument storagedir or the option parallelMap.default.storagedir.

Value

Nothing.

[Package parallelMap version 1.1 Index]