Parallel versions of apply-family functions.

parallelLapply: A parallel lapply() version.
parallelSapply: A parallel sapply() version.
All functions are simple wrappers for parallelMap().

parallelLapply(xs, fun, ..., impute.error = NULL, level = NA_character_)

parallelSapply(
  xs,
  fun,
  ...,
  simplify = TRUE,
  use.names = TRUE,
  impute.error = NULL,
  level = NA_character_
)

Arguments

xs
(vector | list)
fun is applied to the elements of this argument.

fun
function
Function to map over xs.

...
(any)
Further arguments passed to fun.

impute.error
(NULL | function(x))
See parallelMap().

level
(character(1))
See parallelMap().

simplify
(logical(1))
See sapply(). Default is TRUE.

use.names
(logical(1))
See sapply(). Default is TRUE.

Value

For parallelLapply a named list, for parallelSapply it depends on the return value of fun and the settings of simplify and use.names.