parallelMap {parallelMap}	R Documentation

Maps a function over lists or vectors in parallel.

Description

Use the parallelization mode and the other options specified in parallelStart.

Usage

parallelMap(fun, ..., more.args = list(), simplify = FALSE,
  use.names = FALSE, level = as.character(NA), show.info = NA)

Arguments

fun
[function]
Function to map over ....

...
[any]
Arguments to vectorize over (list or vector).

more.args
[list]
A list of other arguments passed to fun. Default is empty list.

simplify
[logical(1)]
Should the result be simplified? See sapply. Default is FALSE.

use.names
[logical(1)]
Should result be named by first vector if that is of class character? Default is FALSE.

level
[character(1)]
If a (non-missing) level is specified in parallelStart, this call is only parallelized if the level specified here matches. Useful if this function is used in a package. Default is NA.

show.info
[logical(1)]
Verbose output on console? Can be used to override setting from options / parallelStart. Default is NA which means no overriding.

Value

Result.

Examples

parallelStart()
parallelMap(identity, 1:2)

## [[1]]
## [1] 1
## 
## [[2]]
## [1] 2

parallelStop()

[Package parallelMap version 1.1 Index]